Refactor "canRunPod" as a soft admit handler
